+++ pdoc/doc.py
@@ -41,10 +41,12 @@ from typing import Generic
 from typing import TypeAlias
 from typing import TypedDict
 from typing import TypeVar
+from typing import cast
 from typing import get_origin
 from typing import is_typeddict
 import warnings
 
+from pdoc import _pydantic
 from pdoc import doc_ast
 from pdoc import doc_pyi
 from pdoc import extract
@@ -209,7 +211,10 @@ class Doc(Generic[T]):
         )
 
 
-class Namespace(Doc[T], metaclass=ABCMeta):
+U = TypeVar("U", bound=types.ModuleType | type)
+
+
+class Namespace(Doc[U], metaclass=ABCMeta):
     """
     A documentation object that can have children. In other words, either a module or a class.
     """
@@ -318,13 +323,17 @@ class Namespace(Doc[T], metaclass=ABCMeta):
                     qualname,
                     docstring="",
                     annotation=self._var_annotations.get(name, empty),
-                    default_value=obj,
+                    default_value=_pydantic.default_value(self.obj, name, obj),
                     taken_from=taken_from,
                 )
-            if self._var_docstrings.get(name):
+
+            if _doc := _pydantic.get_field_docstring(cast(type, self.obj), name):
+                doc.docstring = _doc
+            elif self._var_docstrings.get(name):
                 doc.docstring = self._var_docstrings[name]
-            if self._func_docstrings.get(name) and not doc.docstring:
+            elif self._func_docstrings.get(name) and not doc.docstring:
                 doc.docstring = self._func_docstrings[name]
+
             members[doc.name] = doc
 
         if isinstance(self, Module):
@@ -774,6 +783,12 @@ class Class(Namespace[type]):
         for cls in self._bases:
             sorted, unsorted = doc_ast.sort_by_source(cls, sorted, unsorted)
         sorted.update(unsorted)
+
+        if _pydantic.is_pydantic_model(self.obj):
+            sorted = {
+                k: v for k, v in sorted.items() if k not in _pydantic.IGNORED_FIELDS
+            }
+
         return sorted
 
     @cached_property
@@ -904,7 +919,7 @@ class Function(Doc[types.FunctionType]):
         else:
             unwrapped = func
         super().__init__(modulename, qualname, unwrapped, taken_from)
-        self.wrapped = func
+        self.wrapped = func  # type: ignore
 
     @cache
     @_include_fullname_in_traceback