+++ CHANGELOG.md
@@ -17,6 +17,9 @@
   upstream](https://github.com/trentm/python-markdown2)
 - Add support for keyword args for Google flavor docs.
   ([#840](https://github.com/mitmproxy/pdoc/pull/840), @aleksslitvinovs)
+- Add support for Pydantic-style field docstrings,
+  e.g. `pydantic.Field(description="...")`
+  ([#802](https://github.com/mitmproxy/pdoc/pull/802), @jinnovation)
 
 ## 2025-06-04: pdoc 15.0.4